Compiler warnings
git-svn-id: file:///home/michael/svn/yuni/trunk@1614 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/dynamiclibrary/file.cpp
@@ -176,7 +176,7 @@ namespace DynamicLibrary
 		if (filename)
 		{
 			// Loading
-			const unsigned int length = strlen(filename);
+			const unsigned int length = (unsigned int) strlen(filename);
 			if (!length)
 				return false;
 			wchar_t* buffer = new wchar_t[length + 10];
